Separate old and new tabs completely
