+++ apps/portals-example/src/scss/advance-view.scss
@@ -3,10 +3,6 @@
     @apply border-neutrals-500 w-[40px] h-[24px];
   }
 
-  &-button-cross-dataset-mode {
-    @apply border-hues-400 w-[32px] h-[32px] p-0 ml-4;
-  }
-
   &-filters {
     @apply bg-neutrals-100 py-4 px-6;
