+++ apps/portals-example/src/scss/styles.scss
@@ -11,6 +11,7 @@
 @import 'calendar.scss';
 @import 'attachments-container.scss';
 @import 'dataset-tabs.scss';
+@import 'dataset-tabs-cross-dataset-mode.scss';
 @import 'messages.scss';
 @import 'conversation-list.scss';
 @import 'share.scss';
